requirements.txt
scikit-learn
pillow
gradio
streamlit
numpy
pyinstaller